Felddistanzen: 4D-Array jeweils mit Distanzen von jedem zu jedem Feld implementiert

+++ src/de/fh/stud/Suchen/Suchkomponenten/Knoten.java
@@ -188,7 +188,7 @@ public class Knoten {
         posY = (byte) pos.y;
     }
 
-    public int getCost() {
+    public short getCost() {
         return cost;
     }
